Target Finding and Movement

+++ Assets/Scripts/Weapon.cs
@@ -1,4 +1,4 @@
 public class Weapon : Item
 {
-    public double Attack;
+    public float Attack;
 }
